refactor: controller only manifolds
We need to split the controller only manifolds out for dqlite only. This
will go away, but for now we need this so we can build and run tests.
